chore: fix all golangci-lint v2 issues
- Add explicit int32 casts to fix G115 integer overflow warnings (7)
- Add package comments to fix revive package-comments (2)
- Remove embedded Time field selector to fix QF1008 (1)

+++ gists.go
@@ -1,3 +1,4 @@
+// Package main provides a template-driven markdown generator using GitHub, RSS, and other data sources.
 package main
 
 import (
@@ -25,7 +26,7 @@ func gists(count int) []Gist {
 	var gists []Gist
 	variables := map[string]interface{}{
 		"username": githubv4.String(username),
-		"count":    githubv4.Int(count),
+		"count":    githubv4.Int(int32(count)),
 	}
 	err := gitHubClient.Query(context.Background(), &gistsQuery, variables)
 	if err != nil {
